refactor the caching
